+++ changelog.md
@@ -30,6 +30,7 @@
 
 # Unreleased
 
+- Add new command ``/zdrawer convert <type>`` Convert data to another storage type (zdrawer.convert)
 - Add WorldGuard Access Verification
 - Change default configuration to SQLITE
 - Fixed item deletion error when drawer adds item and fills drawer